backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
numpy==1.26.3
opencv-python==4.9.0.80
Pillow==10.2.0
tensorflow==2.15.0
scikit-image==0.22.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
